[8] Allow images to load for Vim plugin
Instead of writing a tmp file for the HTML, we put a normal file into
the user's home directory. Doing so, Quick Look allows us to access any
file path (e.g. for images) on the system. Assuming the source Markdown
file is also somewhere in the tree below the user's home directory.

close #8

Co-authored-by: pmeinhardt <mnhrdt@gmail.com>